SleepycatSequence to support long batch size
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@12597 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ code/base/common/src/com/tc/util/sequence/MutableSequence.java
@@ -8,7 +8,7 @@ public interface MutableSequence extends Sequence {
   // This provide a unique id for an instance of the sequence.
   public String getUID();
 
-  public long nextBatch(int batchSize);
+  public long nextBatch(long batchSize);
   
   public void setNext(long next);
